requirements.txt
Django==5.0.3
djangorestframework==3.14.0
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
httpx==0.27.0
Pillow==10.2.0
django-environ==0.11.2
gunicorn==21.2.0
python-multipart==0.0.9
django-cors-headers==4.3.1
python-dotenv==1.0.1
dj-database-url==2.1.0